pyproject.toml
[project]
name = "jarvis-v3"
version = "0.2.0"
description = "Voice interface for Claude Code — Deepgram STT + Kokoro MLX TTS"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=5.0",
    "mlx-audio[tts]>=0.3",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "soundfile>=0.12",
    "pyyaml>=6",
    "onnxruntime>=1.17",
    "scipy>=1.12",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
jarvis = "jarvis.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["jarvis"]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
